pyproject.toml
[project]
name = "cep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.124.2",
    "platformdirs>=4.5.1",
    "psutil>=7.2.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "typer>=0.20.0",
    "result", #last version as this is not maintained anymore
]

[project.scripts]
cep = "cep.cli.main:app"

[tool.setuptools.package-data]
cep = ["templates/*.yaml", "app_store/app_templates/*.yaml"]
